Update async_client.py
Fix bugs when have multiple batch object. The problem is previously, no new instance of `batch` object will be return. This will cause problem down the line when use batch object in multiple sequence

To avoid this every time the `batch()` function is called, it will create a new instance of batch object, but still point to the same data source as a single source of truth (Previously it copied to create a new data source for async client)